+++ src/service/database/types.ts
@@ -1,4 +1,5 @@
 import { EngineModel } from "../engine/model";
+import { EngineModelInterface } from "../engine/types";
 
 export type Database = {
   id?: string;
@@ -8,7 +9,7 @@ export type Database = {
 
 export type CreateDatabaseOptions = {
   region?: string;
-  attached_engines?: string[] | EngineModel[];
+  attached_engines?: string[] | EngineModelInterface[];
   description?: string;
   fail_if_exists?: boolean;
 };
@@ -17,9 +18,11 @@ export interface DatabaseModelInterface {
   name: string;
   description: string;
   delete(): Promise<void>;
+  getAttachedEngines(): Promise<EngineModelInterface[]>;
 }
 
 export interface DatabaseServiceInterface {
+  getById(databaseId: string): Promise<DatabaseModelInterface>;
   getByName(name: string): Promise<DatabaseModelInterface>;
   getAll(): Promise<DatabaseModelInterface[]>;
   create(
